requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pandas==2.1.4
numpy==1.24.4
scikit-learn==1.3.2
tensorflow==2.15.0
requests==2.31.0
aiohttp==3.9.1
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1
httpx==0.25.2
factory-boy==3.3.0
soccerdata>=1.5.0